bring "delete pref rep workflow" into "mirror CR show and edit tables" (#534)
https://cits.artic.edu/issues/3151

* switch to CR mutli select

* turn ActiveTriples::Relation into array before taking difference

* 2352_add row param to Solr calls to decrease logging

* adds delete-asset class to delete button on asset show page

* render shared partial if this partial is for preferred reps

* add class to make js bind to click event

* add controller action to return json

* adds handler for click event

* adds route for new relationships endpoint

* adds modal, need to finalize

* configures delete button to function normally if no pref-reps exist

* add pref label to json response

* cleans up modal

* not going to render this based on whether or not asset is a preferred on
the show page load b/c it's strong if the modal markup is on every asset
show page so that if a preferred is added/exists on an old page version,
then the modal will still load since it's AJAX'ing the endpooint which
is checking solr

* easily trigger a click on the delete button again in order to spark the
handler on the element's event. this may not be the best possible
solution since this is mixing js and html, so maybe in the future we can
write more js to select this element and register a short handler that
will trigger the js again if clicked

* add the modal markup to this partial that is loaded on both the "my
assets' page and the 'assets shared with me' tab

* destroy all relationships before destroying the GenericWork.

improve the relationships endpoint to include the preferred CR's edit
path so that it can be rendered via the js

* wipe out the rows of the tbody before adding new ones so we don't get
stale duplicates

also no reason to add ANOTHER confirm box if we are continuing above if
modal_not_needed is true, the data-confirm on the element will take care
of it.

* render shared modal in asset show page

* override so that we can add the target class and data-asset-id for the js

* don't use backticks b/c it breaks stuff:

ExecJS::RuntimeError: SyntaxError: Unexpected character '`' (line: 56907, col: 19, pos: 1839299)

* remove current_user's dept from sharing dropdown

* fix generic work controller destroy tests since asset without pref rep
can be toasted now

* improve guidance text on modal

* forgot to iterate through all CR's; also improved guidance text